Add use-aria-live package for quick examples
Motivation: Quickly develop example stories for developing the addon

Resources: [use-aria-live](https://www.npmjs.com/package/use-aria-live)

Notes:
`use-aria-live` is a snipped of code that I've used to control Aria Live
Regions in other applications. I published it for easy access in today's
assignment but it is not yet documented.